requirements.txt
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
tenacity>=8.2.0
ratelimit>=2.2.1
pyyaml>=6.0
ta>=0.11.0
numpy>=1.24.0
yfinance>=0.2.0
polygon-api-client>=1.0.0
